[RM] Rename files from `smoke_detector_*` to `sd_*` for easier handling
Unfortunately, MCUXpresso has fixed width tabs, and they are too short
for the long names. Thus, rename files so they are distinguishable in
the IDE's tab band.
