+++ misc/Rauchmelder-bcu1/inc/sd_com.h
@@ -20,7 +20,7 @@
 #include <optional>
 #include <sblib/timeout.h>
 
-#include "smoke_detector_shared_enums.h"
+#include "sd_shared_enums.h"
 
 /**
  * Command byte to send to the smoke detector
