+++ misc/Rauchmelder-bcu1/inc/sd_device.h
@@ -22,8 +22,8 @@
 #include <sblib/io_pin_names.h>
 #include <sblib/timeout.h>
 
-#include "smoke_detector_com.h"
-#include "smoke_detector_shared_enums.h"
+#include "sd_com.h"
+#include "sd_shared_enums.h"
 
 class SmokeDetectorAlarm;
 class SmokeDetectorConfig;
