+++ misc/Rauchmelder-bcu1/inc/sd_errorcode.h
