+++ misc/Rauchmelder-bcu1/inc/sd_shared_enums.h
