+++ misc/Rauchmelder-bcu1/src/sd_alarm.cpp
@@ -15,9 +15,9 @@
  *  published by the Free Software Foundation.
  */
 
-#include "smoke_detector_alarm.h"
-#include "smoke_detector_config.h"
-#include "smoke_detector_group_objects.h"
+#include "sd_alarm.h"
+#include "sd_config.h"
+#include "sd_group_objects.h"
 
 SmokeDetectorAlarm::SmokeDetectorAlarm(const SmokeDetectorConfig *config, const SmokeDetectorGroupObjects *groupObjects)
     : config(config),
